+++ src/base/ElementLink.php
@@ -77,7 +77,7 @@ abstract class ElementLink extends Link
     {
         if(is_null($this->_element))
         {
-            $this->_element = Craft::$app->getElements()->getElementById((int) $this->value);
+            $this->_element = Craft::$app->getElements()->getElementById((int) $this->value, static::elementType(), $this->ownerElement->siteId ?? null);
         }
         return $this->_element;
     }